make some apps and news instances
