+++ source/news_instance.html
@@ -1,0 +1,14 @@
+{% extends 'base.html' %}
+{% block content %}
+  <ul class="news__list">
+    <li class="news__item">
+      <h3 class="news__item-title">{{ news.title }}</h3>
+      <p class="news__item-date">{{ news.public_date|date:'j.m.y' }}</p>
+      <img class="news__item-image" src='{{ news.image.url }}' alt="Ректор" width="260"
+           height="195">
+      <p>{{ news.main_text }}</p>
+      <p class="news__item-subtitle">{{ news.preview }}</p>
+      <a class="news__item-link" href="{{ news.get_absolute_url }}">Читать</a>
+    </li>
+  </ul>
+{% endblock %}
