+++ user/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from user import views
+
+urlpatterns = [
+  path('login/', views.login, name='login'),
+  path('register/', views.register, name='register'),
+  path('logout/', views.logout, name='logout')
+]
