+++ web/admin.py
@@ -1,5 +1,7 @@
 from django.contrib import admin
-from web.models import User, Feedback, News
+from web.models import Feedback
+from news.models import News
+from user.models import User
 
 # Register your models here.
 admin.site.register(User)
