+++ webapp/settings.py
@@ -45,6 +45,8 @@ INSTALLED_APPS = [
   'django.contrib.messages',
   'django.contrib.staticfiles',
   'web.apps.WebConfig',
+  'user.apps.UserConfig',
+  'news.apps.NewsConfig',
 ]
 
 MIDDLEWARE = [
