+++ webapp/urls.py
@@ -30,12 +30,10 @@ urlpatterns = [
                 path('admin/', admin.site.urls),
                 path('general/', include('web.urls'), name='general'),
                 path('', RedirectView.as_view(url='/general/', permanent=False)),
-                path('register/', views.register, name='register'),
-                path('login/', views.login, name='login'),
-                path('news/', views.news, name='news'),
-                path('calendar', views.calendar, name='calendar'),
+                path('news/', include('news.urls')),
+                path('calendar/', views.calendar, name='calendar'),
                 path('events', views.events, name='events'),
-                path('logout/', views.logout, name='logout'),
+                path('user/', include('user.urls')),
                 re_path(r'^media/(?P<path>.*)$', serve, {'document_root': settings.MEDIA_ROOT}),
                 re_path(r'^static/(?P<path>.*)$', serve, {'document_root': settings.STATIC_ROOT}),
               ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT) + static(settings.STATIC_URL,
